+++ goodreads/src/main/java/project/goodreads/services/UserService.java
@@ -1,12 +1,11 @@
 package project.goodreads.services;
 
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import project.goodreads.enums.Role;
+import project.goodreads.exceptions.NullException;
 import project.goodreads.exceptions.UserAlreadyExistException;
 import project.goodreads.models.User;
 import project.goodreads.repositories.BookshelfRepository;
@@ -18,17 +17,19 @@ import project.goodreads.repositories.UserRepository;
 public class UserService {
 
     final private UserRepository userRepository;
-    final private PasswordEncoder passwordEncoder;
     final private BookshelfService bookshelfService;
     final private BookshelfRepository bookshelfRepository;
 
     public User getUser(Long id) {
+        if (id == null)
+            throw new NullException("User id cannot be null");
+
         User user = userRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
 
         return user;
     }
 
-    public User createUser(String username, String password, Role role) {
+    public User createUser(String username) {
 
         if (usernameExist(username)) {
             throw new UserAlreadyExistException("User with this username already exists: " + username);
@@ -36,20 +37,13 @@ public class UserService {
 
         User user = new User();
         user.setUsername(username);
-        user.setPassword(passwordEncoder.encode(password));
-        user.setRole(role);
 
-        userRepository.save(user);
+        User savedUser = userRepository.save(user);
 
-        bookshelfService.createBookshelf("read", user, true);
-        bookshelfService.createBookshelf("to be read", user, true);
+        bookshelfService.createBookshelf("read", savedUser);
+        bookshelfService.createBookshelf("to be read", savedUser);
 
         return user;
-    }
-
-    public User createUser(String username, String password) {
-
-        return createUser(username, password, Role.USER);
     }
 
     private boolean usernameExist(String username) {
@@ -57,31 +51,25 @@ public class UserService {
         return userRepository.findByUsername(username).isPresent();
     }
 
-    public User updateUser(User user, String username, String password) {
+    @SuppressWarnings("null")
+    public User updateUser(User user, String username) {
 
         if (username != null && !username.isEmpty())
             user.setUsername(username);
-        if (password != null && !password.isEmpty())
-            user.setPassword(passwordEncoder.encode(password));
 
         userRepository.save(user);
 
         return user;
     }
 
-    public User updateUser(User user, String username, String password, Role role) {
-
-        user.setRole(role);
-
-        return updateUser(user, username, password);
-    }
-
     public void deleteUser(User user) {
 
-        deleteUserById(user.getId());
+        deleteUser(user.getId());
     }
 
-    public void deleteUserById(Long id) {
+    public void deleteUser(Long id) {
+        if (id == null)
+            throw new NullException("User id cannot be null");
 
         bookshelfRepository.deleteAllByUserId(id);
         userRepository.deleteById(id);
